Moved BTTasks to separate folder and setup melee target objective and parent class
