+++ Source/ProjectInfestation/AI/BTTasks/BTTask_FindPatrolPathPoint.cpp
@@ -3,8 +3,8 @@
 
 #include "BTTask_FindPatrolPathPoint.h"
 
-#include "../EnemyCharacter.h"
-#include "EnemyAIController.h"
+#include "../../EnemyCharacter.h"
+#include "../EnemyAIController.h"
 #include "BehaviorTree/BlackboardComponent.h"
 
 UBTTask_FindPatrolPathPoint::UBTTask_FindPatrolPathPoint(FObjectInitializer const& objectInitializer)
