+++ Source/ProjectInfestation/AI/Objectives/AIObjective.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Engine/DataAsset.h"
+#include "AIObjective.generated.h"
+
+/**
+ * 
+ */
+UCLASS(BlueprintType)
+class PROJECTINFESTATION_API UAIObjective : public UDataAsset
+{
+	GENERATED_BODY()
+	
+};
