+++ Source/ProjectInfestation/AI/Objectives/AIObjective_MeleeTarget.cpp
@@ -1,0 +1,5 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "AIObjective_MeleeTarget.h"
+
